requirements.txt
opencv-python
requests
numpy
python-dotenv
simpleaudio

torch
torchvision